+++ Assets/Scripts/Core/GameLoopManager.cs
@@ -1,7 +1,10 @@
+using System;
 using System.Collections;
 using System.Collections.Generic;
+using System.Linq;
 using TMPro;
 using UnityEngine;
+using UnityEngine.Purchasing;
 using UnityEngine.UI;
 
 public class GameLoopManager : MonoBehaviour
@@ -16,8 +19,6 @@ public class GameLoopManager : MonoBehaviour
     public CardUI selectedCard;
     public ManaManager manaManager;
     [SerializeField] GameObject settings;
-    public DeckData selectedDeck;
-
 
     public bool isGameRunning = false;
     private List<GameObject> playerUnits = new();
@@ -30,7 +31,17 @@ public class GameLoopManager : MonoBehaviour
 
     void Start()
     {
-        LevelInformationFadeTextManager.instance.DisplayTextWithFade(GameManager.instance.currentLevelData.currentlevel.ToString(), "Facile");
+        VoidMapGeneratorGPU.instance.SetSeed(VoidMapGeneratorGPU.instance.seed);
+
+        LevelInformationFadeTextManager.instance.DisplayTextWithFade(
+            GameManager.instance.currentLevelData.currentlevel.ToString(),
+            GameManager.instance.currentLevelData.spawnAlgo.difficulty
+        );
+
+        if (GameModeManager.selectedDeck == null)
+            DeckSelectionManager.instance.OpenSelection();
+        else
+            DeckSelectionManager.instance.CloseSelection();
     }
 
     void Update()
@@ -89,7 +100,7 @@ public class GameLoopManager : MonoBehaviour
         }
     }
 
-    private void EndGame(bool isPlayerVictorious)
+    public void EndGame(bool isPlayerVictorious)
     {
         isGameRunning = false;
 
@@ -155,8 +166,11 @@ public class GameLoopManager : MonoBehaviour
         ManaManager.instance.AddMana(3);
     }
 
-    public void StartOrStopCombat(bool enabled = true)
+    public void StartOrStopCombat(bool enabled)
     {
+        if (!playerUnits.Any() || isGameRunning)
+            enabled = !enabled;
+
         isGameRunning = enabled;
 
         GameObject[] entities = GameObject.FindGameObjectsWithTag("Entities");