+++ Assets/Scripts/Core/LevelSelectionManager.cs
@@ -3,37 +3,26 @@ using UnityEngine.UI;
 
 public class LevelSelectManager : MonoBehaviour
 {
-    public GameObject levelButtonPrefab;
-    public int levelCount = 50;
+    [SerializeField] GameObject levelButtonPrefab;
+    [SerializeField] Transform parentsPosition;
 
     private LevelData[] levels;
 
     void Start()
     {
-        LoadLevelsConfig();
         GenerateLevels();
-    }
-
-    void LoadLevelsConfig()
-    {
-        TextAsset jsonFile = Resources.Load<TextAsset>("Levels/levels_config");
-
-        if (jsonFile != null) {
-            LevelsWrapper wrapper = JsonUtility.FromJson<LevelsWrapper>(jsonFile.text);
-            levels = wrapper.levels;
-            GameManager.instance.alllevelData = levels;
-        } else
-            BugTracker.Critical("Failed to load levels configs from 'levels_config.json'.");
     }
 
     void GenerateLevels()
     {
+        levels = GameManager.instance.alllevelData;
+
         for (int i = 1; i <= levels.Length; i++) {
-            GameObject btn = Instantiate(levelButtonPrefab, transform);
+            GameObject btn = Instantiate(levelButtonPrefab, parentsPosition);
 
             RectTransform rt = btn.GetComponent<RectTransform>();
 
-            rt.anchoredPosition = new Vector2((i - 1) * 200f - 600, 0f); // TODO delete -600
+            rt.anchoredPosition = new Vector2((i - 1) * 200f, 0f);
 
             LevelButtonUI ui = btn.GetComponent<LevelButtonUI>();
 
@@ -42,5 +31,11 @@ public class LevelSelectManager : MonoBehaviour
             Button button = btn.GetComponent<Button>();
             button.onClick.AddListener(ui.OnClick);
         }
+    }
+
+    public void UpdateLevels()
+    {
+        levels = GameManager.instance.alllevelData;
+        GenerateLevels();
     }
 }