+++ Assets/Scripts/SceneSpecific/Menu/Deck/DeckSelectionManager.cs
@@ -3,36 +3,35 @@ using System.Collections.Generic;
 
 public class DeckSelectionManager : MonoBehaviour
 {
+    public static DeckSelectionManager instance;
+
     [Header("UI")]
-    public GameObject selectionPanel;
     public Transform decksContainer;
     public GameObject deckPreviewPrefab;
 
     private List<DeckData> availableDecks = new();
 
-    void Start()
+    void Awake()
     {
-        if (selectionPanel != null)
-            selectionPanel.SetActive(false);
+        instance = this;
     }
 
     void LoadDecks()
     {
         DeckData[] loaded = Resources.LoadAll<DeckData>("Decks");
         availableDecks = new List<DeckData>(loaded);
-        BugTracker.Info($"[DeckSelection] {availableDecks.Count} decks chargés.");
+        BugTracker.Info($"[DeckSelection] {availableDecks.Count} decks load.");
     }
 
     public void OpenSelection()
     {
         LoadDecks();
-        selectionPanel.SetActive(true);
         RefreshDecks();
     }
 
     public void CloseSelection()
     {
-        selectionPanel.SetActive(false);
+        gameObject.SetActive(false);
     }
 
     void RefreshDecks()
@@ -40,8 +39,7 @@ public class DeckSelectionManager : MonoBehaviour
         foreach (Transform child in decksContainer)
             Destroy(child.gameObject);
 
-        foreach (DeckData deck in availableDecks)
-        {
+        foreach (DeckData deck in availableDecks) {
             GameObject obj = Instantiate(deckPreviewPrefab, decksContainer);
             DeckPreview preview = obj.GetComponent<DeckPreview>();
             preview.Setup(deck, this);
@@ -51,7 +49,13 @@ public class DeckSelectionManager : MonoBehaviour
     public void SelectDeck(DeckData deck)
     {
         GameModeManager.selectedDeck = deck;
-        BugTracker.Info($"[DeckSelection] Deck '{deck.deckName}' sélectionné.");
+
+        if (DeckManager.instance != null) {
+            DeckManager.instance.SetDeck();
+            ShopManager.instance.LoadCards();
+        }
+
+        BugTracker.Info($"[DeckSelection] Deck '{deck.deckName}' selected.");
         CloseSelection();
     }
 }